cmactions: Use Preference Fragments
PreferenceActivity.addPreferencesFromResource() is deprecated.

Change-Id: I18461cb9c92a22631457942cfcc3737adbf75582
cmactions: Integrate with Settings

Bring this up to speed on the Settings changes in N

Change-Id: I9365b3ebd1bbfed2936302e30da50e3f9af06665
cmactions: Hold a wakelock when the screen is on

Sometimes the device is suspended before ACTION_SCREEN_OFF is received.
When this happens, the proximity sensor doesn't get enabled, breaking
the wake gesture. Hold a wakelock when the screen is on and release it
only once the sensor has been enabled.

Change-Id: I0af8eb9cfb264adb7de628b7b0908eb8933f997e